+++ backends/aws_kinesis/aws_kinesis_put_record.cc
@@ -10,18 +10,18 @@
 
 using namespace Aws;
 
-SDKOptions options;
+static SDKOptions options;
 
-Kinesis::KinesisClient *client;
+static Kinesis::KinesisClient *client;
 
 struct request_outcome {
     Kinesis::Model::PutRecordOutcomeCallable future_outcome;
     size_t data_len;
 };
 
-Vector<request_outcome> request_outcomes;
+static Vector<request_outcome> request_outcomes;
 
-void kinesis_init(const char *region, const char *access_key_id, const char *secret_key, const long timeout) {
+void backends_kinesis_init(const char *region, const char *access_key_id, const char *secret_key, const long timeout) {
     InitAPI(options);
 
     Client::ClientConfiguration config;
@@ -37,13 +37,13 @@ void kinesis_init(const char *region, const char *access_key_id, const char *sec
     }
 }
 
-void kinesis_shutdown() {
+void backends_kinesis_shutdown() {
     Delete(client);
 
     ShutdownAPI(options);
 }
 
-int kinesis_put_record(const char *stream_name, const char *partition_key,
+int backends_kinesis_put_record(const char *stream_name, const char *partition_key,
                        const char *data, size_t data_len) {
     Kinesis::Model::PutRecordRequest request;
 
@@ -56,7 +56,7 @@ int kinesis_put_record(const char *stream_name, const char *partition_key,
     return 0;
 }
 
-int kinesis_get_result(char *error_message, size_t *sent_bytes, size_t *lost_bytes) {
+int backends_kinesis_get_result(char *error_message, size_t *sent_bytes, size_t *lost_bytes) {
     Kinesis::Model::PutRecordOutcome outcome;
     *sent_bytes = 0;
     *lost_bytes = 0;