+++ exporting/graphite/graphite.c
@@ -3,23 +3,6 @@
 #include "graphite.h"
 
 /**
- * Initialize Graphite connector
- *
- * @param instance a connector data structure.
- * @return Always returns 0.
- */
-int init_graphite_connector(struct connector *connector)
-{
-    connector->worker = simple_connector_worker;
-
-    struct simple_connector_config *connector_specific_config = mallocz(sizeof(struct simple_connector_config));
-    connector->config.connector_specific_config = (void *)connector_specific_config;
-    connector_specific_config->default_port = 2003;
-
-    return 0;
-}
-
-/**
  * Initialize Graphite connector instance
  *
  * @param instance an instance data structure.
@@ -27,6 +10,12 @@ int init_graphite_connector(struct connector *connector)
  */
 int init_graphite_instance(struct instance *instance)
 {
+    instance->worker = simple_connector_worker;
+
+    struct simple_connector_config *connector_specific_config = mallocz(sizeof(struct simple_connector_config));
+    instance->config.connector_specific_config = (void *)connector_specific_config;
+    connector_specific_config->default_port = 2003;
+
     instance->start_batch_formatting = NULL;
     instance->start_host_formatting = format_host_labels_graphite_plaintext;
     instance->start_chart_formatting = NULL;
@@ -115,7 +104,7 @@ int format_host_labels_graphite_plaintext(struct instance *instance, RRDHOST *ho
  */
 int format_dimension_collected_graphite_plaintext(struct instance *instance, RRDDIM *rd)
 {
-    struct engine *engine = instance->connector->engine;
+    struct engine *engine = instance->engine;
     RRDSET *st = rd->rrdset;
     RRDHOST *host = st->rrdhost;
 
@@ -156,7 +145,7 @@ int format_dimension_collected_graphite_plaintext(struct instance *instance, RRD
  */
 int format_dimension_stored_graphite_plaintext(struct instance *instance, RRDDIM *rd)
 {
-    struct engine *engine = instance->connector->engine;
+    struct engine *engine = instance->engine;
     RRDSET *st = rd->rrdset;
     RRDHOST *host = st->rrdhost;